added score to json file

+++ backend/src/CrimeSum.java
@@ -2,9 +2,9 @@ import java.util.Vector;
 
 public class CrimeSum {
 	Vector<Crime> crimeSet;
-	int score;
+	double score;
 	
-	CrimeSum(Vector<Crime> crimeSet, int score){
+	CrimeSum(Vector<Crime> crimeSet, double score){
 		this.crimeSet = crimeSet;
 		this.score = score;
 	}
